# UniversityApp/swagger.yaml
openapi: 3.0.1
info:
  title: University API
  description: This is a university api documentation
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /group/details:
    get:
      operationId: showDetails
      parameters:
      - name: groupId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /course/delete:
    get:
      operationId: deleteCourse
      parameters:
      - name: courseId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/teacher/{teacherId}/addCourse/{courseId}:
    put:
      tags:
      - Teacher
      summary: "Adds course by = {courseId} to teacher by id = {teacherId}"
      operationId: addCourse
      parameters:
      - name: teacherId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: courseId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /teacher/deleteCourse:
    get:
      operationId: deleteCourseFromTeacher
      parameters:
      - name: teacherId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: courseId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /lecture/details:
    get:
      operationId: showDetails_1
      parameters:
      - name: lectureId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/student/searchByFirstName/{firstName}:
    get:
      tags:
      - Student
      summary: Finds students by first name
      operationId: searchByFirstName
      parameters:
      - name: firstName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
...
  /group/deleteFromFaculty:
    get:
      operationId: deleteFromFaculty
      parameters:
      - name: groupId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: facultyId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /student/searchByFirstName:
    get:
      operationId: searchByFirstName_1
      parameters:
      - name: firstName
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /faculty/details:
    get:
      operationId: showDetails_2
      parameters:
      - name: facultyId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
  /teacher/searchByLastName:
    get:
      operationId: searchByLastName
      parameters:
      - name: lastName
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /teacher/formForUpdate:
    get:
      operationId: updateTeacher
      parameters:
      - name: teacherId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /teacher/delete:
    get:
      operationId: deleteTeacher
      parameters:
      - name: teacherId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/group/{groupId}/deleteFromFaculty:
    delete:
      tags:
      - Group
      summary: Deletes group from its faculty
      operationId: deleteFromFaculty_1
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
  /teacher/deleteFromFaculty:
    get:
      operationId: deleteFromFaculty_2
      parameters:
      - name: teacherId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: facultyId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /teacher/searchByFirstName:
    get:
      operationId: searchByFirstName_2
      parameters:
      - name: firstName
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /teacher/addCourse:
    post:
      operationId: addCourseToTeacher
      parameters:
      - name: teacherId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CrmObject'
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/teacher/searchByLastName/{lastName}:
    get:
      tags:
      - Teacher
      summary: Finds teachers by last name
      operationId: searchByLastName_1
      parameters:
      - name: lastName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
  /api/teacher/{teacherId}/fromFaculty/{facultyId}:
    delete:
      tags:
      - Teacher
      summary: "Deletes teacher by id = {teacherId} from faculty by id = {facultyId}"
      operationId: deleteFromFaculty_3
      parameters:
      - name: teacherId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: facultyId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/student/searchByLastName/{lastName}:
    get:
      tags:
      - Student
      summary: Finds students by last name
      operationId: searchByLastName_2
      parameters:
      - name: lastName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
...
  /teacher/details:
    get:
      operationId: showDetails_3
      parameters:
      - name: teacherId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /teacher/saveTeacher:
    post:
      operationId: saveTeacher
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/teacher/searchByFirstName/{firstName}:
    get:
      tags:
      - Teacher
      summary: Finds teachers by first name
      operationId: searchByFirstName_3
      parameters:
      - name: firstName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
...
  /student/searchByLastName:
    get:
      operationId: searchByLastName_3
      parameters:
      - name: lastName
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
  /api/teacher/{teacherId}/course/{courseId}:
    delete:
      tags:
      - Teacher
      summary: "Deletes course by id = {courseId} from teacher by id = {teacherId}"
      operationId: deleteCourseFromTeacher_1
      parameters:
      - name: teacherId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: courseId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
...
    CrmObject:
      required:
      - id
      type: object
      properties:
        id:
          minimum: 1
          type: integer
          format: int32
...
    Student:
      required:
      - email
      - firstName
      - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          maxLength: 2147483647
          minLength: 2
          type: string
        lastName:
          maxLength: 2147483647
          minLength: 2
          type: string
        email:
          type: string
...
    Teacher:
      required:
      - firstName
      - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          maxLength: 2147483647
          minLength: 2
          type: string
        lastName:
          maxLength: 2147483647
          minLength: 2
          type: string
...